chatbotrequirements.txt
sounddevice
keyboard
openai-whisper
langchain
scipy
elevenlabs